refactor: 從 main.js 抽出 capacityStatus、connectionStatus、pdfExport 模組
main.js 1716 → 1569 行（-147 行）
- updateCapacityStatus → src/utils/capacityStatus.js
- checkConnectionStatus → src/utils/connectionStatus.js
- printSchedule / exportToPdf → src/features/schedule/pdfExport.js
